Assert: Term
