+++ FL_TESTS
@@ -4,6 +4,7 @@ test/AssertDefuzzifier.h
 test/AssertFactory.h
 test/AssertHedge.h
 test/AssertNorm.h
+test/AssertTerm.h
 test/Headers.h
 test/Listener.h
 test/Mock.h
